AuthServer: Separated server startup to new shared helpers
