+++ ChatModApp.AuthCallback/Program.cs
@@ -1,30 +1,5 @@
-using Microsoft.Fast.Components.FluentUI;
+using ChatModApp.AuthCallback;
 
-var builder = WebApplication.CreateBuilder(args);
+var builder = StartupHelpers.CreateBuilder(new() {Args = args});
 
-// Add services to the container.
-builder.Services.AddRazorPages();
-builder.Services.AddServerSideBlazor();
-builder.Services.AddFluentUIComponents();
-builder.Services.AddHttpClient();
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (!app.Environment.IsDevelopment())
-{
-    app.UseExceptionHandler("/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-    app.UseHsts();
-}
-
-app.UseHttpsRedirection();
-
-app.UseStaticFiles();
-
-app.UseRouting();
-
-app.MapBlazorHub();
-app.MapFallbackToPage("/_Host");
-
-app.Run();
+await StartupHelpers.CreateApplication(builder).RunAsync();
