+++ ChatModApp.AuthCallback/Startup.cs
@@ -1,0 +1,43 @@
+using Microsoft.Fast.Components.FluentUI;
+
+namespace ChatModApp.AuthCallback;
+
+public static class StartupHelpers
+{
+    public static WebApplicationBuilder CreateBuilder(WebApplicationOptions options)
+    {
+        var builder = WebApplication.CreateBuilder(options);
+
+        // Add services to the container.
+        builder.Services.AddRazorPages();
+        builder.Services.AddServerSideBlazor();
+        builder.Services.AddFluentUIComponents();
+        builder.Services.AddHttpClient();
+
+        return builder;
+    }
+
+    public static WebApplication CreateApplication(WebApplicationBuilder builder)
+    {
+        var app = builder.Build();
+
+        // Configure the HTTP request pipeline.
+        if (!app.Environment.IsDevelopment())
+        {
+            app.UseExceptionHandler("/Error");
+            // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+            app.UseHsts();
+        }
+
+        app.UseHttpsRedirection();
+
+        app.UseStaticFiles();
+
+        app.UseRouting();
+
+        app.MapBlazorHub();
+        app.MapFallbackToPage("/_Host");
+
+        return app;
+    }
+}
